requirements.txt
tweepy==4.14.0
anthropic>=0.25.0
requests==2.31.0
python-dotenv==1.0.0
